Move clr_module
